Break into different components
